allow auto-remove of credential exchange records to be disabled; track credential revocation registry and ID; update performance demo
