+++ aries_cloudagent/revocation/error.py
@@ -1,0 +1,11 @@
+"""Revocation error classes."""
+
+from ..core.error import BaseError
+
+
+class RevocationError(BaseError):
+    """Base exception for revocation-related errors."""
+
+
+class RevocationNotSupportedError(RevocationError):
+    """Attempted to create registry for non-revocable cred def."""
